this is my life and yes it is worth it!!
Whatever the study, whatever the outcome, know that behavior is learned. We as humans do nothing on instinct. That being said I will not allow this simulated sett back to change what I know to be reality . I am not so weak minded that I would let anything or any one convince me of what I already know to be true